Unify all basic dice rolling under one command - it used to be five!
