+++ package.json
@@ -26,7 +26,7 @@
     "@dice-roller/rpg-dice-roller": "^5.1.0",
     "@discordjs/builders": "^0.13.0",
     "@sapphire/discord.js-utilities": "^4.11.1",
-    "@sapphire/framework": "^3.0.0-next.b7485a2.0",
+    "@sapphire/framework": "^3.0.0-next.080c438.0",
     "@sapphire/plugin-logger": "^2.2.1",
     "bufferutil": "^4.0.5",
     "discord.js": "^13.7.0",
