+++ src/commands/dice/roll.ts
@@ -1,5 +1,5 @@
 import { DiceRoll } from "@dice-roller/rpg-dice-roller";
-import { Command } from "@sapphire/framework";
+import { Command, RegisterBehavior } from "@sapphire/framework";
 import type { CommandInteraction } from "discord.js";
 
 export default class RollCommand extends Command {
@@ -23,21 +23,65 @@ export default class RollCommand extends Command {
               .setName("dice")
               .setDescription("The dice to roll")
               .setRequired(true)
+          )
+          .addIntegerOption((option) =>
+            option
+              .setName("repetitions")
+              .setDescription("The number of times to roll the dice")
+              .setRequired(false)
+          )
+          .addStringOption((option) =>
+            option
+              .setName("output")
+              .setDescription(
+                "Display the average, minimum, or maximum possible value for the provided dice"
+              )
+              .addChoices(
+                { name: "average", value: "averageTotal" },
+                { name: "minimum", value: "minTotal" },
+                { name: "maximum", value: "maxTotal" }
+              )
+              .setRequired(false)
           ),
-      { guildIds: ["513474679583801394"] }
+      {
+        guildIds: ["513474679583801394"],
+        behaviorWhenNotIdentical: RegisterBehavior.Overwrite,
+      }
     );
   }
 
-  public chatInputRun(interaction: CommandInteraction) {
+  public async chatInputRun(interaction: CommandInteraction) {
     const notation = interaction.options.getString("dice");
+    const repetitions = interaction.options.getInteger("repetitions");
+    const display = interaction.options.getString("output") || "output";
 
     if (notation) {
+      if (repetitions) {
+        const rolls: DiceRoll[] = [];
+
+        for (let i = 0; i < repetitions; i++) {
+          const roll = new DiceRoll(notation);
+          rolls.push(roll);
+        }
+
+        const author = interaction.member?.toString();
+        const replies = rolls.map((r, i) => {
+          if (i === 0) {
+            return `${author}, ${r[display as keyof typeof r]}`;
+          }
+          return `${r[display as keyof typeof r]}`;
+        });
+        const followups = replies.slice(1);
+
+        await interaction.reply(replies[0]);
+        for (const reply of followups) {
+          await interaction.followUp(reply);
+        }
+        return;
+      }
       const roll = new DiceRoll(notation);
       const author = interaction.member?.toString();
-      const reply =
-        roll.output.length > 2000
-          ? `${author}, ${roll.notation} = ${roll.total}`
-          : `${author}, ${roll.output}`;
+      const reply = `${author}, ${roll[display as keyof typeof roll]}`;
       return interaction.reply(reply);
     }
   }
